drop likes table only if exists

+++ db/migrate/20170517064830_drop_likes_table.rb
@@ -1,6 +1,8 @@
 class DropLikesTable < ActiveRecord::Migration[5.0]
   def up
-    drop_table :likes
+    if ActiveRecord::Base.connection.data_source_exists? 'likes'
+      drop_table :likes
+    end
   end
 
   def down
